requirements.txt
matplotlib>=3.0
requests>=2.28.0
beautifulsoup4>=4.11.0

Flask>=2.2.0

PyJWT>=2.8.0
cryptography>=41.0.0
PyGithub>=2.2.0
